pyproject.toml
[project]
name = "docker-reconciler"
version = "0.1.0"
description = "Stub reconciler"
dependencies = [
    "pydantic",
    "tomli; python_version < '3.11'",
]

[project.scripts]
reconciler = "src.main:main"
test-dev-env = "scripts.test_dev_env:main"
run-e2e-container-test = "runbook.run_e2e_container_test:main"

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
    "lizard",
    "import-linter",
    "pre-commit",
    "pytest",
    "ansible",
    "ansible-lint",
    "pyyaml",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.importlinter]
root_package = "src"

[[tool.importlinter.contracts]]
name = "Reconciler must not import utils"
type = "forbidden"
source_modules = ["src.reconciler"]
forbidden_modules = ["src.utils"]

[[tool.importlinter.contracts]]
name = "Utils must not import reconciler"
type = "forbidden"
source_modules = ["src.utils"]
forbidden_modules = ["src.reconciler"]

[[tool.importlinter.contracts]]
name = "Models must not import anything internal"
type = "forbidden"
source_modules = ["src.models"]
forbidden_modules = ["src.reconciler", "src.utils"]
